Fix overflow error and make filters usable
Change schemas.json to use long instead of int for host memory field,
avoiding this error:

    com.fasterxml.jackson.databind.JsonMappingException: Numeric value (50638880768) out of range of int

Replace the use of class Filters with Map<String,String> in service
templates, avoiding this error when calling list(filter) methods:

    java.lang.IllegalArgumentException: Map must include generic types

+++ src/main/java/io/rancher/type/Host.java
@@ -66,7 +66,7 @@ public class Host extends AbstractType {
     
     private Integer localStorageMb;
     
-    private Integer memory;
+    private Long memory;
     
     private Integer milliCpu;
     
@@ -334,11 +334,11 @@ public class Host extends AbstractType {
       this.localStorageMb = localStorageMb;
     }
     
-    public Integer getMemory() {
+    public Long getMemory() {
         return this.memory;
     }
 
-    public void setMemory(Integer memory) {
+    public void setMemory(Long memory) {
       this.memory = memory;
     }
